deleted unused @ResponseStatus, add location in create AdminRestaurantController, changed MenuItem.price from Double to Integer, replaced boot type EAGER to LAZY

+++ src/main/java/ru/didorenko/votingsystem/model/MenuItem.java
@@ -31,7 +31,7 @@ public class MenuItem extends NamedEntity {
 
     @Column(name = "price")
     @NotNull
-    private Double price;
+    private Integer price; // Price in cents for better precision
 
     @Column(name = "menu_item_date")
     @NotNull
